fix sync observation barrier and share test mocks
