+++ glue/src/stateful/actor/core/syncing.rs
@@ -291,236 +291,28 @@ mod tests {
     use super::Syncing;
     use crate::stateful::{
         actor::syncer::{self, SyncResult},
-        db::{Anchor, AttachableResolver, ManagedDb, Merkleized, Unmerkleized},
-        Application, Proposed,
+        db::{Anchor, AttachableResolver},
+        tests::mocks::{anchor, test_databases, TestApp, TestBlock, TestScheme, TestVariant},
     };
     use commonware_actor::mailbox as actor_mailbox;
-    use commonware_codec::{EncodeSize, Error as CodecError, Read, ReadExt as _, Write};
     use commonware_consensus::{
-        marshal::{self, core::Actor as MarshalActor, standard::Standard},
-        simplex::{mocks::scheme as scheme_mocks, types::Context as SimplexContext},
-        types::{Epoch, FixedEpocher, Height, View, ViewDelta},
-        Block as ConsensusBlock, CertifiableBlock, Heightable,
+        marshal::{self, core::Actor as MarshalActor},
+        simplex::mocks::scheme as scheme_mocks,
+        types::{FixedEpocher, ViewDelta},
+        Heightable,
     };
-    use commonware_cryptography::{
-        certificate::ConstantProvider, ed25519, sha256::Digest as Sha256Digest, Digest as _,
-        Digestible, Signer as _,
-    };
+    use commonware_cryptography::{certificate::ConstantProvider, sha256::Digest as Sha256Digest};
     use commonware_parallel::Sequential;
     use commonware_runtime::{
-        buffer::paged::CacheRef, deterministic, Buf, BufMut, ContextCell, Runner as _,
-        Supervisor as _,
+        buffer::paged::CacheRef, deterministic, ContextCell, Runner as _, Supervisor as _,
     };
     use commonware_storage::archive::immutable;
     use commonware_utils::{
         acknowledgement::Exact, channel::oneshot, sync::AsyncRwLock, Acknowledgement, NZUsize,
         NZU16, NZU64,
     };
-    use futures::{FutureExt, Stream};
-    use std::{convert::Infallible, sync::Arc};
-
-    type TestDatabases = Arc<AsyncRwLock<TestDb>>;
-    type TestVariant = Standard<TestBlock>;
-    type TestScheme = scheme_mocks::Scheme<ed25519::PublicKey>;
-
-    #[derive(Clone, Copy)]
-    struct TestUnmerkleized;
-
-    #[derive(Clone, Copy)]
-    struct TestMerkleized;
-
-    impl Unmerkleized for TestUnmerkleized {
-        type Merkleized = TestMerkleized;
-        type Error = Infallible;
-
-        async fn merkleize(self) -> Result<Self::Merkleized, Self::Error> {
-            Ok(TestMerkleized)
-        }
-    }
-
-    impl Merkleized for TestMerkleized {
-        type Digest = Sha256Digest;
-        type Unmerkleized = TestUnmerkleized;
-
-        fn root(&self) -> Self::Digest {
-            Sha256Digest::from([0; 32])
-        }
-
-        fn new_batch(&self) -> Self::Unmerkleized {
-            TestUnmerkleized
-        }
-    }
-
-    #[derive(Default)]
-    struct TestDb;
-
-    impl<E: Send> ManagedDb<E> for TestDb {
-        type Unmerkleized = TestUnmerkleized;
-        type Merkleized = TestMerkleized;
-        type Error = Infallible;
-        type Config = ();
-        type SyncTarget = u64;
-
-        async fn init(_context: E, _config: Self::Config) -> Result<Self, Self::Error> {
-            Ok(Self)
-        }
-
-        async fn new_batch(_db: &Arc<AsyncRwLock<Self>>) -> Self::Unmerkleized {
-            TestUnmerkleized
-        }
-
-        fn matches_sync_target(_batch: &Self::Merkleized, _target: &Self::SyncTarget) -> bool {
-            true
-        }
-
-        async fn finalize(&mut self, _batch: Self::Merkleized) -> Result<(), Self::Error> {
-            Ok(())
-        }
-
-        async fn sync_target(&self) -> Self::SyncTarget {
-            0
-        }
-
-        async fn rewind_to_target(&mut self, _target: Self::SyncTarget) -> Result<(), Self::Error> {
-            Ok(())
-        }
-    }
-
-    #[derive(Clone, Debug, PartialEq, Eq)]
-    struct TestBlock {
-        context: SimplexContext<Sha256Digest, ed25519::PublicKey>,
-        height: Height,
-        digest: Sha256Digest,
-    }
-
-    impl TestBlock {
-        fn new(height: u64, digest_byte: u8) -> Self {
-            Self {
-                context: SimplexContext {
-                    round: commonware_consensus::types::Round::new(
-                        Epoch::zero(),
-                        View::new(height),
-                    ),
-                    leader: ed25519::PrivateKey::from_seed(0).public_key(),
-                    parent: (View::zero(), Sha256Digest::EMPTY),
-                },
-                height: Height::new(height),
-                digest: Sha256Digest::from([digest_byte; 32]),
-            }
-        }
-    }
-
-    impl Write for TestBlock {
-        fn write(&self, buf: &mut impl BufMut) {
-            self.context.write(buf);
-            buf.put_u64(self.height.get());
-            buf.put_slice(self.digest.as_ref());
-        }
-    }
-
-    impl EncodeSize for TestBlock {
-        fn encode_size(&self) -> usize {
-            self.context.encode_size() + 8 + 32
-        }
-    }
-
-    impl Read for TestBlock {
-        type Cfg = ();
-
-        fn read_cfg(buf: &mut impl Buf, _: &Self::Cfg) -> Result<Self, CodecError> {
-            let context = SimplexContext::read(buf)?;
-            let height = Height::new(buf.get_u64());
-            let mut digest = [0u8; 32];
-            buf.copy_to_slice(&mut digest);
-            Ok(Self {
-                context,
-                height,
-                digest: Sha256Digest::from(digest),
-            })
-        }
-    }
-
-    impl Digestible for TestBlock {
-        type Digest = Sha256Digest;
-
-        fn digest(&self) -> Self::Digest {
-            self.digest
-        }
-    }
-
-    impl Heightable for TestBlock {
-        fn height(&self) -> Height {
-            self.height
-        }
-    }
-
-    impl ConsensusBlock for TestBlock {
-        fn parent(&self) -> Self::Digest {
-            Sha256Digest::EMPTY
-        }
-    }
-
-    impl CertifiableBlock for TestBlock {
-        type Context = SimplexContext<Sha256Digest, ed25519::PublicKey>;
-
-        fn context(&self) -> Self::Context {
-            self.context.clone()
-        }
-    }
-
-    #[derive(Clone)]
-    struct TestApp;
-
-    impl Application<deterministic::Context> for TestApp {
-        type SigningScheme = TestScheme;
-        type Context = SimplexContext<Sha256Digest, ed25519::PublicKey>;
-        type Block = TestBlock;
-        type Databases = TestDatabases;
-        type InputProvider = ();
-
-        fn sync_targets(
-            block: &Self::Block,
-        ) -> <Self::Databases as crate::stateful::db::DatabaseSet<deterministic::Context>>::SyncTargets{
-            block.height().get()
-        }
-
-        async fn genesis(&mut self) -> Self::Block {
-            TestBlock::new(0, 0)
-        }
-
-        async fn propose(
-            &mut self,
-            _context: (deterministic::Context, Self::Context),
-            _ancestry: impl Stream<Item = Self::Block> + Send,
-            _batches: <Self::Databases as crate::stateful::db::DatabaseSet<
-                deterministic::Context,
-            >>::Unmerkleized,
-            _input: &mut Self::InputProvider,
-        ) -> Option<Proposed<Self, deterministic::Context>> {
-            None
-        }
-
-        async fn verify(
-            &mut self,
-            _context: (deterministic::Context, Self::Context),
-            _ancestry: impl Stream<Item = Self::Block> + Send,
-            _batches: <Self::Databases as crate::stateful::db::DatabaseSet<deterministic::Context>>::Unmerkleized,
-        ) -> Option<<Self::Databases as crate::stateful::db::DatabaseSet<deterministic::Context>>::Merkleized>{
-            None
-        }
-
-        async fn apply(
-            &mut self,
-            _context: (deterministic::Context, Self::Context),
-            _block: &Self::Block,
-            _batches: <Self::Databases as crate::stateful::db::DatabaseSet<
-                deterministic::Context,
-            >>::Unmerkleized,
-        ) -> <Self::Databases as crate::stateful::db::DatabaseSet<deterministic::Context>>::Merkleized
-        {
-            TestMerkleized
-        }
-    }
+    use futures::FutureExt;
+    use std::sync::Arc;
 
     #[derive(Clone)]
     struct NoopResolver;
@@ -553,7 +345,7 @@ mod tests {
                     held_verify_requests: Vec::new(),
                     database_subscribers: Vec::new(),
                     artifact: Some(SyncResult {
-                        databases: Arc::new(AsyncRwLock::new(TestDb)),
+                        databases: test_databases(),
                         anchor,
                     }),
                     resolvers: NoopResolver,
@@ -628,14 +420,6 @@ mod tests {
         )
         .await;
         mailbox
-    }
-
-    fn anchor(height: u64, digest_byte: u8) -> Anchor<Sha256Digest> {
-        Anchor {
-            height: Height::new(height),
-            round: commonware_consensus::types::Round::new(Epoch::zero(), View::new(height)),
-            digest: Sha256Digest::from([digest_byte; 32]),
-        }
     }
 
     #[test]
