+++ glue/src/stateful/tests/mod.rs
@@ -18,6 +18,7 @@ use single_db_app::SingleDbEngine;
 use std::time::Duration;
 
 mod common;
+pub(crate) mod mocks;
 mod multi_db_app;
 mod properties;
 mod single_db_app;
